Implemented dynamic loading of platforms from DLLs.  Also fixed a lot of compilation errors and warnings under Windows.

+++ platforms/cuda/tests/TestCudaBrownianIntegrator.cpp
@@ -121,7 +121,7 @@ void testTemperature() {
     }
     pe /= steps;
     double expected = 0.5*numBonds*BOLTZ*temp;
-    ASSERT_EQUAL_TOL(expected, pe, 20*expected/std::sqrt(steps));
+    ASSERT_EQUAL_TOL(expected, pe, 20*expected/std::sqrt((double) steps));
 }
 
 void testConstraints() {
